moved button back, and close works
