+++ packages/molly/src/lib/components/Molly.svelte
@@ -76,8 +76,6 @@
 </script>
 
 <MollyButton>
-	<div class="chat-container fixed bottom-0 right-0 w-80 bg-slate-200 dark:bg-slate-600">
-		<MollyHeader />
 			<div class="h-full grid grid-rows-[1fr_auto]">
 					<div bind:this={chatWindow} class="messages-container h-80 bg-surface-500/30 overflow-y-auto">
 						{#each chatMessages as message}
@@ -97,5 +95,4 @@
 					}}
 					/>
 			</div>
-	</div>
 </MollyButton>
